Added basic layout and table for admin to home
